unsignedness fixes
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4178 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/ui_basic/table.cc
@@ -168,7 +168,7 @@ void Table<void *>::draw(RenderTarget & dst)
 	dst.brighten_rect(Rect(Point(0, 0), get_w(), get_h()), ms_darken_value);
 
 	while (idx < m_entry_records.size()) {
-		if (y >= get_h())
+		if (y >= static_cast<int32_t>(get_h()))
 			return;
 
 		const Entry_Record & er = *m_entry_records[idx];
